requirements.txt
fastapi 
uvicorn
pydantic
openai
streamlit
python-dotenv
autogen-agentchat
autogen-ext[openai]
PyPDF2
requests==2.31.0
aiofiles==23.2.1